+++ README.md
@@ -9,14 +9,12 @@ React is a front-end JavaScript library designed for building scalable and dynam
 
 React is versatile and isn’t limited to web apps. It can also be used for creating mobile and desktop apps through platforms like React Native for mobile development and Electron for desktop applications. While this course focuses on web development, the concepts apply across these different environments.
 
-## First step (First day)
+## First step
 
 [setting-up-your-environment](https://www.bitovi.com/academy/learn-react/setting-up-your-environment.html)
 
 Branch: [01-setting-up-your-environment](https://shields.io/)
 
-cd place-my-order
-npm install
 
 ## Second step (First day)
 
@@ -24,12 +22,44 @@ npm install
 
 Branch: [02-introduction-to-jsx](https://github.com/thomazbandeira/bitovi-academy-learn-core-react/tree/02-introduction-to-jsx)
 
-Exercise 01
+#### Exercise 01
+
 copy broken test and implement html to JSX
 
-Exercise 02
+#### Exercise 02
+
 01 - Update the existing JSX to render the list of restaurants
+
 02 - Use Array.map() to iterate over the restaurants.data
+
 03 - Make sure to use key inside the .map()
+
 04 - Render <p>No restaurants.</p> if, hypothetically, there weren’t any restaurants.
 
+## third step (First day)
+
+[Building Components](https://www.bitovi.com/academy/learn-react/components.html)
+
+Branch: [03-building-Components](https://github.com/thomazbandeira/bitovi-academy-learn-core-react/blob/03-building-components)
+
+
+#### SETUP
+
+ Create src/pages/ (folder)
+
+ Create src/pages/RestaurantList/ (folder)
+
+ Create src/pages/RestaurantList/index.ts
+
+ Create src/pages/RestaurantList/RestaurantList.tsx
+
+ Update src/App.tsx
+
+ Create src/pages/RestaurantList/RestaurantList.test.tsx
+
+ Update src/App.test.tsx
+
+ #### Exercise 01 
+
+ 01 - Move the logic in our App component to our new RestaurantList component.
+02 - Update our App component to use our new RestaurantList component.
