+++ place-my-order/src/pages/RestaurantList/RestaurantList.tsx
@@ -1,0 +1,7 @@
+import CheeseThumbnail from 'place-my-order-assets/images/2-thumbnail.jpg'
+import PoutineThumbnail from 'place-my-order-assets/images/4-thumbnail.jpg'
+
+const RestaurantList: React.FC = () => {
+}
+
+export default RestaurantList
