+++ place-my-order/src/pages/RestaurantList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./RestaurantList"
